pyproject.toml
[build-system]
requires = ["poetry-core>=2.2.1"]
build-backend = "poetry.core.masonry.api"

[project]
name = "pychub"
version = "2.0.0"
description = "Make a Python-native .chub installer from wheels and their dependencies."
readme = "README.md"
requires-python = ">=3.10,<4.0"
license = { file = "LICENSE" }
authors = [
    { name = "Steve Storck", email = "steve973@gmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Archiving :: Packaging",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Software Distribution",
    "Topic :: Utilities"
]
keywords = [
    "bundle",
    "chub",
    "executable",
    "package",
    "packaging",
    "zipapp"
]

[project.urls]
Homepage = "https://github.com/Steve973/pychub"
Documentation = "https://github.com/Steve973/pychub/blob/main/docs/USER_GUIDE.md"
Changelog = "https://github.com/Steve973/pychub/blob/main/docs/CHANGELOG.md"
ReleaseNotes = "https://github.com/Steve973/pychub/blob/main/docs/RELEASE_NOTES.md"
Contributing = "https://github.com/Steve973/pychub/blob/main/docs/DEV_AND_CONTRIB.md"
Source = "https://github.com/Steve973/pychub/tree/main"

# Plugin links
PoetryPlugin = "https://github.com/Steve973/pychub-build-plugins/tree/main/poetry"
PoetryPluginPyPI = "https://pypi.org/project/pychub-poetry-plugin/"

HatchPlugin = "https://github.com/Steve973/pychub-build-plugins/tree/main/hatch"
HatchPluginPyPI = "https://pypi.org/project/pychub-hatch-plugin/"

PDMPlugin = "https://github.com/Steve973/pychub-build-plugins/tree/main/pdm"
PDMPluginPyPI = "https://pypi.org/project/pychub-pdm-plugin/"

[project.scripts]
pychub = "pychub.package.cli:main"

[tool.poetry]
packages = [{ include = "pychub", from = "src" }]

[tool.poetry.dependencies]
appdirs = ">=1.4.4"
attrs = ">=25.3.0"
cyclonedx-python-lib = ">=11.2.0"
packaging = ">=25.0"
pyyaml = ">=6.0.3"
requests = ">=2.32.5"
#spdx-tools = ">=0.8.3"
tomli = ">=2.3.0"
tomli-w = ">=1.2.0"

[tool.poetry.group.dev.dependencies]
build = ">=1.3.0"
mypy = ">=1.18.2"
poetry = ">=2.2.1"
pytest = ">=8.4.1"
pytest-cov = ">=7.0.0"
ruff = ">=0.14.5"

[tool.pytest.ini_options]
addopts = ["-ra", "-q", "--import-mode=importlib"]
testpaths = ["."]
python_files = ["test_*.py"]
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')"
]

[tool.ruff]
# Treat these as your source roots
src = ["src", "tests"]

line-length = 100
target-version = "py310"

[tool.ruff.lint]
# Start conservative, adjust as you like
select = ["E", "F", "W"]
ignore = []

[tool.mypy]
python_version = "3.10"
files = ["src"]
namespace_packages = true
strict_optional = true
pretty = true
show_error_codes = true

# Tune as you go
warn_unused_ignores = true
warn_return_any = true
check_untyped_defs = true

